# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Redirect Service API
  description: |
    Servico de redirecionamento HTTP latency-critical com tracking de clicks e regras condicionais.

    ## Visao geral
    - **Hot path**: `GET /`, `GET /db`, `GET /:slug` retornam 302 navegacional para o melhor link disponivel (ranking eCPM/RPS).
    - **API administrativa**: `/api/*` para gerenciar grupos de dominio, links, rules e in-app rules.
    - **Health**: `/health`, `/health/detailed`, `/health/ready`, `/ping` (Kubernetes-friendly).

    ## Conceitos
    - **Domain group**: agrupamento logico de dominios (ex: `main`, `db`, ou slugs custom). Cada grupo tem ranking eCPM proprio (recalculado a cada 15 min via cron).
    - **Rule**: regra condicional que faz match em query params (ex: `utm_source=facebook&country=BR`) e redireciona para uma URL fixa.
    - **In-app rule**: regra baseada em `utm_campaign`. Se o user-agent for navegador in-app (Facebook/Instagram), redireciona; caso contrario, devolve HTML com iframe fullscreen para o destino.
    - **Broad clicks**: contador agregado por `broad` query param + data (YYYY-MM-DD).
  version: 1.0.0
  contact:
    name: Caio Norder
    email: caio@joinads.me
servers:
  - url: https://redirect.belnk.in
    description: Producao
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /db:
    get:
      tags: [redirect]
      summary: Redirect para o grupo `db`
      description: Mesma logica de `/` mas usa o ranking do grupo `db` em vez de `main`.
      parameters:
        - $ref: '#/components/parameters/UtmCampaign'
        - $ref: '#/components/parameters/UtmSource'
        - $ref: '#/components/parameters/UtmMedium'
        - $ref: '#/components/parameters/Broad'
      responses:
        '200':
          description: HTML com iframe (InApp rule + nao-in-app UA)
          content:
            text/html:
              schema:
                type: string
        '204':
          description: Favicon
        '302':
          description: Redirect para o destino
        '503':
          description: Grupo `db` sem dominios configurados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /db/{campaignId}:
    get:
      tags: [redirect]
      summary: Redirect grupo `db` com campaignId no path
      description: Alias de `/db?utm_campaign={campaignId}`. O `campaignId` do path e usado como `utm_campaign` para match em `InAppRules`.
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
          description: Equivalente a `utm_campaign`
      responses:
        '200':
          description: HTML com iframe
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirect
...
  /api/domain-groups:
    get:
      tags: [domain-groups]
      summary: Listar todos os grupos de dominio
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items: { $ref: '#/components/schemas/DomainGroup' }
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [domain-groups]
      summary: Criar grupo de dominio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                  description: Identificador (lowercase, numeros, hifens)
                  example: campaign-summer
                name:
                  type: string
                  example: Campaign Summer 2026
                domains:
                  type: array
                  items: { type: string, format: hostname }
                  example: ['site1.com', 'site2.com']
            examples:
              minimal:
                value:
                  slug: campaign-summer
                  name: Campaign Summer 2026
              with-domains:
                value:
                  slug: news
                  name: News Network
                  domains: ['noticia.com', 'jornal.net']
      responses:
        '201':
          description: Grupo criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: '#/components/schemas/DomainGroup' }
        '400':
          description: slug ou name ausente, ou slug invalido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Slug ja existe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/InternalError'
  /api/domain-groups/{slug}:
    put:
      tags: [domain-groups]
      summary: Atualizar grupo (slug, name, bestRpsMode)
      parameters:
        - $ref: '#/components/parameters/GroupSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Pelo menos um campo deve ser informado
              properties:
                slug:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                  description: Novo slug (renomeia o grupo)
                name: { type: string }
                bestRpsMode:
                  type: boolean
                  description: |
                    Se true, ativa modo best-RPS (busca o link de melhor RPS cujo dominio o IP ainda nao visitou nesta hora).
                    Se false, usa round-robin global por counter Redis.
            examples:
              rename:
                value: { slug: novo-slug, name: Novo Nome }
              toggle-mode:
                value: { bestRpsMode: true }
      responses:
        '200':
          description: Grupo atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: '#/components/schemas/DomainGroup' }
        '400':
          description: Nenhum campo informado ou slug invalido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflito - novo slug ja existe
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [domain-groups]
      summary: Deletar grupo
      description: O grupo `main` nao pode ser deletado (HTTP 403).
      parameters:
        - $ref: '#/components/parameters/GroupSlug'
      responses:
        '200':
          description: Deletado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Group 'campaign-summer' deleted" }
        '403':
          description: Tentativa de deletar grupo `main`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    GroupSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9-]+$'
      description: Slug do grupo de dominio
      example: main
...
    UtmCampaign:
      name: utm_campaign
      in: query
      schema: { type: string }
      description: Identificador de campanha. Tambem usado para match em `InAppRules`.
    UtmSource:
      name: utm_source
      in: query
      schema: { type: string }
      description: Origem do trafego. Default `redron` se ausente.
    UtmMedium:
      name: utm_medium
      in: query
      schema: { type: string }
      description: Meio do trafego. Default `broadcast` se ausente.
    Broad:
      name: broad
      in: query
      schema: { type: string }
      description: Quando presente, usa como `utm_campaign` E incrementa contador `broad_clicks` no Mongo.
...
  responses:
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Erro interno
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: 'Failed to do something' }
        message: { type: string, example: 'Optional details' }
        timestamp: { type: string, format: date-time }
      required: [error]
...
    DomainGroup:
      type: object
      description: Grupo de dominios. Cada grupo tem ranking proprio e configuracao de modo (best-RPS ou round-robin).
      properties:
        _id: { type: string, description: ObjectId do MongoDB }
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          example: main
        name:
          type: string
          example: Main Domains
        domains:
          type: array
          items: { type: string, format: hostname }
          example: ['site1.com', 'site2.com']
        active:
          type: boolean
          example: true
        bestRpsMode:
          type: boolean
          description: Se true, usa best-RPS por IP (mais lento, mais inteligente). Se false/ausente, round-robin.
          example: false
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [slug, name, domains, active]
...